+++ Makefile
@@ -5,7 +5,7 @@ NAME	=	minishell
 include	./includes/make/vars.mk
 
 ##	C FILES
-SRC_FILES	=	main.c startup.c parser.c tokenization.c shutdown.c execute.c pathing.c builtin.c mario.c
+SRC_FILES	=	main.c startup.c parser.c tokenization.c shutdown.c execute.c pathing.c builtin.c builtin_1.c mario.c
 SRC			=	${addprefix ./src/, ${SRC_FILES}}
 OBJ			=	${SRC:.c=.o}
